move two remaining caret functions into single module. Update PM to latest commit (still not working for empty ranges)
