--- document/static/js/es6_modules/editor/collab/carets/mod.js
@@ -1,11 +1,0 @@
-import {ModCollabCaretLayout} from "./layout"
-import {ModCollabCaretInteractions} from "./interactions"
-
-export class ModCollabCarets {
-  constructor(mod) {
-    mod.carets = this
-    this.mod = mod
-    new ModCollabCaretLayout(this)
-    new ModCollabCaretInteractions(this)
-  }
-}
