+++ document/static/js/es6_modules/editor/collab/doc-changes.js
@@ -87,7 +87,7 @@ export class ModCollabDocChanges {
                 diff_version: this.mod.editor.pm.mod.collab.version,
                 diff: toSend.steps.map(s => s.toJSON()),
                 footnote_diff: fnToSend.steps.map(s => s.toJSON()),
-                currentPosition:this.mod.carets.interactions.caretPosition(),
+                currentPosition:this.mod.carets.caretPosition(),
                 comments: this.mod.editor.mod.comments.store.unsentEvents(),
                 comment_version: this.mod.editor.mod.comments.store.version,
                 request_id: request_id,
@@ -135,7 +135,7 @@ export class ModCollabDocChanges {
             this.mod.editor.mod.footnotes.fnEditor.applyDiffs(data.footnote_diff)
         }
         if (data.currentPosition) {
-                this.mod.carets.layout.updateCaret(data.currentPosition)
+                this.mod.carets.updateCaret(data.currentPosition)
         }
         if (data.reject_request_id) {
             this.rejectDiff(data.reject_request_id)
